ci: remove greenkeeper lockfile upload step
The GreenKeeper lockfile upload step, requires having a secure token
injected into the public continuous integration environment.
With the current GitHub oauth API, this would mean the generating user
would need to create a token, that gives write access to ALL of their
public GitHub repositories.
This might be feasible with a build user scoped to a single repository,
or if GitHub improves their API.
However at the moment this seems like a security risk.

Short term the best solution appears to be disabling the upload step, so
CI doesn't waste time attempting a push that will not work.