fix(package): require npm 5.6.0 or higher
npm <5.6.0 has issues with optional dependencies that are only installed
for  particular operating system.
causing dependencies such as fsevents to be added to the
package-lock.json file on MacOS, but then be removed when a Linux system
installs the project.
npm >=5.6.0 includes a fix that normalizes handling of optional
dependencies to allow for any OS to produce that same package-lock.json
file from a given package.json