fix(list-of-links): multiple links when first redundant with launch bar (#855)
list-of-links has a feature where when it's configured with a single link redundant with the widget launch bar, it collapses to a basic widget as the  preferred way to present a single hyperlink. That feature was bugged so that it would suppress 2-4 link list-of-links widgets when the first link matched the launch bar. This fixes the bug, adding to localhost content a demo of the fixed case.