docs(timesensitive): clarify takeActionEndDate (#861)
* docs(timesensitive): clarify takeActionEndDate

Got confused about how to configure takeActionEndDate such that Benefit Information widget prematurely told employees today was last day to enroll, whereas tomorrow is actual last day to enroll. Updating documentation with clarity that might have supported my correctly configuring the widget.

* docs(timesensitive): prefer 00:00 over 12:00am

It's clearer, with less opportunity to get confused about whether that's noon or midnight and if midnight does that go with the beginning or ending day and oh goodness dates and times are complicated.

Also, 00:00 is closer to what you'd actually write in this configuration to configure a specific time, so it sets the reader a little more along the successful path for configuring this.