build(travis): only use openjdk8
OpenJDK 8 is closest to what MyUW in practice runs in production.  Differential failures depending upon the choice of JDK are rare in this project. Building multiple times on multiple JDKs is slowing down the feedback loop from Travis. It'll be better to build less (slightly less information) and get that resolved feedback faster (sooner clearly unblocked from merging to master).